Rock, Paper, Scissors Game
Made a rock, paper, scissors game that's played in the developer console. Random number generator's play each other and announce a winner. Made a user input version too that prompts you for a number (assigned to either rock, paper, or scissors), then compares the two, and finally announces the winner.